django

+++ database/migrations/2023_04_01_003209_hotels.php
@@ -6,13 +6,13 @@ use Illuminate\Support\Facades\Schema;
 class Hotels extends Migration
 {
     public function up()
-    {
+    {        
         Schema::table('hotels', function (Blueprint $table) {
             // $table->id();
             // $table->timestamps();
             $table->string('name', 255);
-            $table->string('direccion', 255);
-            $table->integer('telefono');
+            $table->string('direccion', 255);            
+            $table->bigInteger('telefono');
             $table->boolean('activo')->default(false);            
         });
     }
